+++ src/GeometryBasics.jl
@@ -8,11 +8,11 @@ module GeometryBasics
     include("fixed_arrays.jl")
     include("offsetintegers.jl")
     include("basic_types.jl")
+    include("primitives.jl")
     include("interfaces.jl")
     include("metadata.jl")
     include("viewtypes.jl")
     include("geometry_primitives.jl")
-    include("rectangles.jl")
     include("meshes.jl")
     include("triangulation.jl")
     include("lines.jl")
