+++ src/primitives.jl
@@ -1,0 +1,8 @@
+# TODO: define common API for all primitives
+
+# include implementations
+include("primitives/rectangles.jl")
+include("primitives/spheres.jl")
+include("primitives/cylinders.jl")
+include("primitives/pyramids.jl")
+include("primitives/particles.jl")
