+++ src/primitives/particles.jl
@@ -1,0 +1,4 @@
+struct Particle{N, T} <: GeometryPrimitive{N, T}
+    position::Point{N, T}
+    velocity::Vec{N, T}
+end
